refactor: collectFields to separate utility
